split types from construction
